refactor(cli): rename stitch.py to stitch_mtp.py
Keep the CLI module filename consistent with the script it replaces.
