+++ src/speculators/cli/stitch_mtp.py
